+++ lib/src/builders/map_json_view_builder.dart
@@ -19,5 +19,5 @@ class MapJsonViewBuilder implements JsonViewBuilder {
   }
 
   @override
-  JsonViewTheme? get jsonViewTheme => _commonBuilder.jsonViewTheme;
+  JsonViewTheme get jsonViewTheme => _commonBuilder.jsonViewTheme;
 }