+++ lib/src/builders/string_json_view_builder.dart
@@ -20,5 +20,5 @@ class StringJsonViewBuilder implements JsonViewBuilder {
   }
 
   @override
-  JsonViewTheme? get jsonViewTheme => _commonBuilder.jsonViewTheme;
+  JsonViewTheme get jsonViewTheme => _commonBuilder.jsonViewTheme;
 }