+++ lib/src/theme/json_view_theme.dart
@@ -1,33 +1,64 @@
 import 'package:flutter/material.dart';
 
+enum JsonViewType {
+  base,
+  collapsible,
+}
+
 class JsonViewTheme {
   const JsonViewTheme({
-    this.keyStyle = const TextStyle(color: Colors.deepPurple),
-    this.doubleStyle = const TextStyle(color: Colors.blue),
-    this.intStyle = const TextStyle(color: Colors.blue),
-    this.boolStyle = const TextStyle(color: Colors.green),
-    this.stringStyle = const TextStyle(color: Colors.amber),
-    this.closeIcon = const Icon(Icons.arrow_drop_up, size: 18),
-    this.openIcon = const Icon(Icons.arrow_drop_down, size: 18),
+    this.defaultTextStyle = const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.w500,
+      color: Colors.white,
+    ),
+    TextStyle? keyStyle,
+    TextStyle? doubleStyle,
+    TextStyle? intStyle,
+    TextStyle? boolStyle,
+    TextStyle? stringStyle,
+    this.closeIcon = const Icon(
+      Icons.arrow_drop_up,
+      size: 18,
+      color: Colors.white,
+    ),
+    this.openIcon = const Icon(
+      Icons.arrow_drop_down,
+      size: 18,
+      color: Colors.white,
+    ),
     this.errorWidget = const Text('error'),
-    this.separator = const Text(' : '),
+    this.separator,
     this.loadingWidget = const CircularProgressIndicator(),
-  });
+    this.viewType = JsonViewType.collapsible,
+    this.backgroundColor = const Color(0xFF1E1F28),
+  })  : _keyStyle = keyStyle ?? const TextStyle(color: Colors.deepPurple),
+        _doubleStyle = doubleStyle ?? const TextStyle(color: Colors.blue),
+        _intStyle = intStyle ?? const TextStyle(color: Colors.blue),
+        _boolStyle = boolStyle ?? const TextStyle(color: Colors.green),
+        _stringStyle = stringStyle ?? const TextStyle(color: Colors.amber);
 
   /// Style of json Object keys name displaying
-  final TextStyle keyStyle;
+  TextStyle get keyStyle => defaultTextStyle.merge(_keyStyle);
+
+  final TextStyle _keyStyle;
 
   /// Style of json Object [double] values displaying
-  final TextStyle doubleStyle;
+  TextStyle get doubleStyle => defaultTextStyle.merge(_doubleStyle);
+
+  final TextStyle _doubleStyle;
 
   /// Style of json Object [int] values displaying
-  final TextStyle intStyle;
+  TextStyle get intStyle => defaultTextStyle.merge(_intStyle);
+  final TextStyle _intStyle;
 
   /// Style of json Object [bool] values displaying
-  final TextStyle boolStyle;
+  TextStyle get boolStyle => defaultTextStyle.merge(_boolStyle);
+  final TextStyle _boolStyle;
 
   /// Style of json Object [String] values displaying
-  final TextStyle stringStyle;
+  TextStyle get stringStyle => defaultTextStyle.merge(_stringStyle);
+  final TextStyle _stringStyle;
 
   /// This icon is shown when the object's fields are hidden
   final Widget closeIcon;
@@ -40,9 +71,19 @@ class JsonViewTheme {
   final Widget errorWidget;
 
   /// Key and value separator widget
-  final Widget separator;
+  final Widget? separator;
 
   /// When loading data from a file, (web in future)
   /// this widget will be shown at the time of processing
   final Widget loadingWidget;
+
+  /// How JsonView widget will build items
+  final JsonViewType viewType;
+
+  /// Background color of JsonView
+  final Color backgroundColor;
+
+  /// This style used as default for all styles
+  /// and for not parsed values
+  final TextStyle defaultTextStyle;
 }